+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 Features:
 
   - Removed dependency on xmlsimple. Processing was so minimal it was pointless.
+
+Bugfixes:
+
+  - Handle files with no extension
   
 ---
 ## 0.1.1 (June 28, 2010)